Refactor: Document 컴포넌트 수정 - DocumentList와 같은 Level에 있도록 수정
